done consultation modul

+++ app/Http/Controllers/Backsite/ConsultationController.php
@@ -32,9 +32,8 @@ class ConsultationController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('consultation_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $consultation = Consultation::orderBy('name', 'ASC')->get();
-
-
         return view('pages.backsite.master-data.consultation.index', compact('consultation'));
     }
 
@@ -56,9 +55,7 @@ class ConsultationController extends Controller
      */
     public function store(StoreConsultationRequest $request)
     {
-        // abort_if(Gate::denies('consultation_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $consultation = Consultation::create($request->all());
-
 
         alert()->success('Success Message', 'Successfully added new Consultation');
         return redirect(route('consultation.index'));
@@ -72,6 +69,7 @@ class ConsultationController extends Controller
      */
     public function show(Consultation $consultation)
     {
+        abort_if(Gate::denies('consultation_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('pages.backsite.master-data.consultation.show', compact('consultation'));
     }
 
@@ -83,6 +81,7 @@ class ConsultationController extends Controller
      */
     public function edit(Consultation $consultation)
     {
+        abort_if(Gate::denies('consultation_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('pages.backsite.master-data.consultation.edit', compact('consultation'));
     }
 
@@ -110,9 +109,9 @@ class ConsultationController extends Controller
      */
     public function destroy(Consultation $consultation)
     {
-        // dd($id);
-        $consultation->forceDelete();
+        abort_if(Gate::denies('consultation_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+        $consultation->forceDelete();
 
         alert()->success('Success Message', 'Successfully deleted Consultation');
         return back();
